requirements.txt
python==3.8.5
django==3.1.1
pillow==7.2.0
